# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"User controller","description":"Public controller. For setting up subscription"},{"name":"Private cover controller","description":"For saving covers & release generation by authenticated users"},{"name":"Public cover controller","description":"For covers generation and regeneration"},{"name":"Public playlist controller","description":"For getting archive"},{"name":"Authentication","description":"Public controller"},{"name":"User controller","description":"Private controller. For changing user's properties"},{"name":"Private playlist controller","description":"For getting playlist from different sources"}],"paths":{"/cover/track/generate":{"post":{"tags":["Private cover controller"],"summary":"Request to generate cover for release","operationId":"createReleaseCover","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReleaseRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReleaseNewDto"}}}}}}},"/cover/playlist/generate":{"post":{"tags":["Public cover controller"],"summary":"Request to generate cover for playlist","operationId":"createPlaylistWithCover","parameters":[{"name":"vibe","in":"query","required":false,"schema":{"type":"string","enum":["DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]}},{"name":"is_abstract","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"is_lofi","in":"query","required":false,"schema":{"type":"boolean","default":true}},{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlDto"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaylistNewDto"}}}}}}},"/auth/sign-up":{"post":{"tags":["Authentication"],"summary":"User's registration","operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"201":{"description":"Created"}}}},"/auth/sign-in":{"post":{"tags":["Authentication"],"summary":"User's authorization","operationId":"signIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponse"}}}}}}},"/user/update/username":{"patch":{"tags":["User controller"],"summary":"Update username","operationId":"updateUsername","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDto"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"202":{"description":"Accepted"}}}},"/playlist/unlike":{"patch":{"tags":["Private playlist controller"],"summary":"unlike playlist","operationId":"unlikePlaylist","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"Playlist_Id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"202":{"description":"Accepted"}}}},"/playlist/like":{"patch":{"tags":["Private playlist controller"],"summary":"like playlist","operationId":"likePlaylist","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"Playlist_Id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"202":{"description":"Accepted"}}}},"/cover/playlist/save":{"patch":{"tags":["Private cover controller"],"summary":"save generated playlist using playlistId and boolean isPrivate","operationId":"savePlaylist","parameters":[{"name":"Playlist_Id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"Cover_Id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"is_private","in":"query","required":true,"schema":{"type":"boolean"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaylistSaveDto"}}}}}}},"/cover/playlist/regenerate":{"patch":{"tags":["Public cover controller"],"summary":"Request to regenerate cover by playlist id","operationId":"updatePlaylistsCover","parameters":[{"name":"vibe","in":"query","required":false,"schema":{"type":"string","enum":["DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]}},{"name":"is_abstract","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"Playlist_Id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}},{"name":"is_lofi","in":"query","required":false,"schema":{"type":"boolean","default":true}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaylistUpdateDto"}}}}}}},"/user/verify/subscription":{"get":{"tags":["User controller"],"summary":"verify subscription","operationId":"verifySubscription","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"202":{"description":"Accepted"}}}},"/user/subscription":{"get":{"tags":["User controller"],"summary":"get current user's subscription info","operationId":"getUserSubscription","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserSubscriptionDto"}}}}}}},"/user/me":{"get":{"tags":["User controller"],"summary":"get current user's profile info","operationId":"getCurrentUserProfile","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileDto"}}}}}}},"/user/find":{"get":{"tags":["User controller"],"summary":"search users in 'find users'","operationId":"searchUsers","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"search","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"202":{"description":"Accepted","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserSmallDto"}}}}}}}},"/playlist/user_collection":{"get":{"tags":["Private playlist controller"],"summary":"get user's playlists by searching his name in 'find users'","operationId":"getUserPlaylists","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"User_Id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"filter","in":"query","required":false,"schema":{"type":"string","enum":["ABSTRACT","NOT_ABSTRACT","LO_FI","HI_FI","DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlaylistUserCollectionDto"}}}}}}}},"/playlist/my_collection":{"get":{"tags":["Private playlist controller"],"summary":"get my playlists","operationId":"getMyPlaylists","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"filter","in":"query","required":false,"schema":{"type":"string","enum":["ABSTRACT","NOT_ABSTRACT","LO_FI","HI_FI","DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlaylistMyCollectionDto"}}}}}}}},"/playlist/archive":{"get":{"tags":["Public playlist controller"],"summary":"get archive of playlists","operationId":"getArchive","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"filter","in":"query","required":false,"schema":{"type":"string","enum":["ABSTRACT","NOT_ABSTRACT","LO_FI","HI_FI","DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlaylistArchiveDto"}}}}}}}},"/cover/track/generate/style":{"get":{"tags":["Private cover controller"],"summary":"Request to get list of 8 styles that describe final cover of release","operationId":"getCoverStyles","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/cover/track/generate/mood":{"get":{"tags":["Private cover controller"],"summary":"Request to get list of 8 adjectives that describe mood of music in release","operationId":"getMusicMoods","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/auth/verify":{"get":{"tags":["Authentication"],"summary":"User's email verification. Used as a link sent to email","operationId":"verifyUser","parameters":[{"name":"arg0","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ErrorResponse":{"type":"object","properties":{"message":{"type":"string"},"reason":{"type":"string"},"status":{"type":"string"},"timestamp":{"type":"string"}}},"ReleaseRequest":{"required":["isLoFi"],"type":"object","properties":{"title":{"maxLength":50,"minLength":1,"type":"string","description":"Title of track/album","example":"Slippery rock"},"mood":{"maxItems":3,"minItems":1,"type":"array","description":"List of adjectives / short phrases","example":"Foggy, Slow, Nostalgic","items":{"type":"string","description":"List of adjectives / short phrases","example":"Foggy, Slow, Nostalgic"}},"object":{"maxLength":255,"minLength":2,"type":"string","description":"Object / action that we see on the cover","example":"wet stone covered with moss, glows a little"},"surrounding":{"maxLength":255,"minLength":2,"type":"string","description":"Description of surrounding of an object or action","example":"forest edge, the sun breaks through the bark of the trees"},"isLoFi":{"type":"boolean","description":"Model of visual AI: dalle-2 or dalle-3"},"coverDescription":{"maxItems":3,"minItems":1,"type":"array","description":"List of adjectives / short phrases that describe cover's style","example":"Blurred, 3D","items":{"type":"string","description":"List of adjectives / short phrases that describe cover's style","example":"Blurred, 3D"}}}},"CoverReleaseDto":{"type":"object","properties":{"link":{"type":"string"},"isLoFi":{"type":"boolean"}}},"ReleaseNewDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"author":{"$ref":"#/components/schemas/UserReleaseDto"},"cover":{"$ref":"#/components/schemas/CoverReleaseDto"},"createdAt":{"type":"string","format":"date-time"}}},"UserReleaseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"email":{"type":"string"},"hiFiReleaseGenerations":{"type":"integer","format":"int32"},"loFiReleaseGenerations":{"type":"integer","format":"int32"}}},"UrlDto":{"required":["link"],"type":"object","properties":{"link":{"maxLength":1000,"minLength":5,"type":"string"}}},"Cover":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"created":{"type":"string","format":"date-time"},"link":{"maxLength":1000,"minLength":5,"type":"string"},"prompt":{"type":"string"},"isAbstract":{"type":"boolean"},"isLoFi":{"type":"boolean"}}},"PlaylistNewDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"url":{"type":"string"},"vibe":{"type":"string","enum":["DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]},"isPrivate":{"type":"boolean"},"isSaved":{"type":"boolean"},"author":{"$ref":"#/components/schemas/UserSmallDto"},"tracks":{"type":"array","items":{"$ref":"#/components/schemas/TrackInPlaylistDto"}},"cover":{"$ref":"#/components/schemas/Cover"},"hiFiGenerationsLeft":{"type":"integer","format":"int32"},"loFiGenerationsLeft":{"type":"integer","format":"int32"}}},"TrackInPlaylistDto":{"type":"object","properties":{"title":{"type":"string"},"authors":{"type":"string"}}},"UserSmallDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"email":{"type":"string"},"loFiPlaylistGenerations":{"type":"integer","format":"int32"},"hiFiPlaylistGenerations":{"type":"integer","format":"int32"}}},"SignUpRequest":{"required":["email","username"],"type":"object","properties":{"username":{"maxLength":30,"minLength":2,"pattern":"^(?=\\S+$).*$","type":"string","description":"Username","example":"Jon"},"email":{"maxLength":254,"minLength":6,"type":"string","description":"User's email address","example":"jondoe@gmail.com"},"password":{"maxLength":27,"minLength":8,"type":"string","description":"Password","example":"Kozyasolov12!"}},"description":"Sign up request"},"SignInRequest":{"required":["email"],"type":"object","properties":{"email":{"maxLength":254,"minLength":6,"type":"string","description":"Email address","example":"jondoe@gmail.com"},"password":{"maxLength":27,"minLength":8,"type":"string","description":"Password","example":"kozyasolov12!"}},"description":"Sign in request"},"JwtAuthenticationResponse":{"type":"object","properties":{"token":{"type":"string","description":"Token","example":"eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTYyMjUwNj..."}},"description":"Response with token"},"UserUpdateDto":{"required":["username"],"type":"object","properties":{"username":{"maxLength":30,"minLength":2,"type":"string"}}},"PlaylistSaveDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"isPrivate":{"type":"boolean"},"author":{"$ref":"#/components/schemas/UserSmallDto"},"savedAt":{"type":"string","format":"date-time"}}},"PlaylistUpdateDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"cover":{"$ref":"#/components/schemas/Cover"},"author":{"$ref":"#/components/schemas/UserSmallDto"},"hiFiGenerationsLeft":{"type":"integer","format":"int32"},"loFiGenerationsLeft":{"type":"integer","format":"int32"}}},"UserSubscriptionDto":{"type":"object","properties":{"email":{"type":"string"},"subscribed":{"type":"boolean"},"generationsLeft":{"type":"integer","format":"int32"},"generationsUpdateAt":{"type":"string","format":"date-time"}}},"UserProfileDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"email":{"type":"string"},"hiFiReleaseGenerations":{"type":"integer","format":"int32"},"loFiReleaseGenerations":{"type":"integer","format":"int32"},"hiFiPlaylistGenerations":{"type":"integer","format":"int32"},"loFiPlaylistGenerations":{"type":"integer","format":"int32"}}},"PlaylistUserCollectionDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"url":{"type":"string"},"vibe":{"type":"string","enum":["DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]},"tracks":{"type":"array","items":{"$ref":"#/components/schemas/TrackInPlaylistDto"}},"cover":{"$ref":"#/components/schemas/Cover"},"isLiked":{"type":"boolean"}}},"PlaylistMyCollectionDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"url":{"type":"string"},"vibe":{"type":"string","enum":["DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]},"isPrivate":{"type":"boolean"},"tracks":{"type":"array","items":{"$ref":"#/components/schemas/TrackInPlaylistDto"}},"cover":{"$ref":"#/components/schemas/Cover"},"isLiked":{"type":"boolean"}}},"PlaylistArchiveDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"url":{"type":"string"},"vibe":{"type":"string","enum":["DANCING_FLOOR","NATURE_DOES_NOT_CARE","BREAKING_DOWN","CAMPFIRE_CALMNESS","TOUGH_AND_STRAIGHT","ENDLESS_JOY"]},"tracks":{"type":"array","items":{"$ref":"#/components/schemas/TrackInPlaylistDto"}},"cover":{"$ref":"#/components/schemas/Cover"},"isLiked":{"type":"boolean"}}}}}}